Register All Necessary Message Listeners for Gateway and Improve Logging (#40)
* Improved logging message clarity

* Refactored ForwardOrderConfirmationMessage

Removed OrderType and OrderStatus from ForwardOrderConfirmationMessage.
This was done to match the changes made to the OrderConfirmationMessage
class. OrderStatus is inferred from the restingQuantity value sent. OrderType
is inferred via orderID once the ForwardOrderConfirmationMessage reaches
the Player.

* Implemented Gateway handlers for the Order Submission Protocol

* Implemented Gateway handlers for the Order Cancellation Protocol

* Cleaned up a lambda in initTopOfBookBroadcast

* Updated unit tests for ForwardCancelConfirmationMessage and
ForwardOrderConfirmationMessage.

Updated unit tests for ForwardCancelConfirmationMessage and
ForwardOrderConfirmationMessage to include tests for respective
CancelConfirmationMessage and OrderConfirmationMessage constructors.

* Renamed initRegisterMatchingEngineListener to initMatchingEnginMessageListeners to make intent more clear